[READ DETAILS] Rethink job system
Ok, so basically, creating jobs that are able keep the players engaged
is something very hard to do using only mods, because something like
that would require messing around with structures and what
not (considering that jobs would work like minigames). So now, the
jobs system works like the following:
The mod will provide a "Moneygiver" block that can be activated with
mesecons, and this block, when activated, will drop money. Creating
the actual jobs and its related minigames is now a job of the player,
and this mod will only provide what's needed to give the players money
after completing a job/minigame.
To make creating minigames easier, mesecons mod is planned to be added
in the future.